RGBGifDisplay/pyproject.toml
[tool.poetry]
name = "RGBGifDisplay"
version = "0.1.0"
description = "Display GIFs on an RGB controller"
authors = [
  "Nicholas Lee <csudcy@gmail.com>"
]

[tool.poetry.dependencies]
python = "^3.7,<3.9"  # Compatible python versions must be declared here
click = "^7.1.2"
Pillow = "^8.0.1"
"luma.emulator" = "^1.3.0"
"luma.led_matrix" = "^1.5.0"


[tool.poetry.dev-dependencies]
isort = "^5.6.4"
pytype = "^2020.10.8"
yapf = "^0.30.0"
